requirements.txt
numpy
scikit-image
pillow
pyautogui
paddleocr
paddlepaddle
